requirements.txt
Flask
flask-jwt-extended
bcrypt
mysql-connector-python
